+++ src/graphics/Screen.cpp
@@ -31,7 +31,7 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #include "graphics/images.h"
 #include "main.h"
 #include "mesh-pb-constants.h"
-#include "meshwifi/meshwifi.h"
+#include "mesh/wifi/WiFiAPClient.h"
 #include "plugins/TextMessagePlugin.h"
 #include "target_specific.h"
 #include "utils.h"
