+++ src/main.cpp
@@ -19,11 +19,10 @@
 #include "concurrency/Periodic.h"
 #include "graphics/Screen.h"
 #include "main.h"
-#include "meshwifi/meshhttp.h"
-#include "meshwifi/meshwifi.h"
-#include "mesh/wifi/WebServerThread.h"
-#include "sleep.h"
+#include "mesh/wifi/WebServer.h"
+#include "mesh/wifi/WiFiAPClient.h"
 #include "plugins/Plugins.h"
+#include "sleep.h"
 #include "target_specific.h"
 #include <OneButton.h>
 #include <Wire.h>
@@ -583,7 +582,7 @@ void loop()
 #endif
 
     // TODO: This should go into a thread handled by FreeRTOS.
-    //handleWebResponse();
+    // handleWebResponse();
 
     service.loop();
 
@@ -596,5 +595,4 @@ void loop()
     // We want to sleep as long as possible here - because it saves power
     mainDelay.delay(delayMsec);
     // if (didWake) DEBUG_MSG("wake!\n");
-
 }
