+++ src/mesh/wifi/ContentHelper.h
@@ -1,0 +1,8 @@
+#include <Arduino.h>
+#include <functional>
+
+
+
+
+void replaceAll(std::string &str, const std::string &from, const std::string &to);
+
